feat: enhance 3D model manipulation with new mesh operations and node implementations - Introduced mesh-ops.ts for advanced operations including transformations and normal recalculation on GLB models. - Added new nodes for loading, saving, and converting 3D models, improving user interaction with model files. - Refactored document-ops.ts to streamline GLB model merging and processing. - Removed unused utility functions to clean up codebase. - Expanded tests to cover new functionalities and ensure robust performance across model operations.
